views and components separated; no need to import navbar and footer;
