+++ src/App.js
@@ -1,18 +1,20 @@
 import React from "react";
 import { BrowserRouter, Route } from "react-router-dom";
-import MentorDashboard from "./components/dashboard/MentorDashboard";
-import StudentDashboard from "./components/dashboard/StudentDashboard";
-import MentorForm from "./components/form/MentorForm";
-import ProjectForm from "./components/form/ProjectForm";
-import StudentForm from "./components/form/StudentForm";
+import Footer from "./components/Footer";
+import Navbar from "./components/Navbar";
 import OAuth from "./components/OAuth";
-import MentorStats from "./components/stats/MentorStats";
-import StudentStats from "./components/stats/StudentStats";
-import ProjectsTable from "./components/tables/ProjectsTable";
-import StudentsTable from "./components/tables/StudentsTable";
 import FAQ from "./views/FAQ";
 import Home from "./views/Home";
+import MentorDashboard from "./views/MentorDashboard";
+import MentorForm from "./views/MentorForm";
+import MentorStats from "./views/MentorStats";
+import ProjectForm from "./views/ProjectForm";
 import Projects from "./views/Projects";
+import ProjectsTable from "./views/ProjectsTable";
+import StudentDashboard from "./views/StudentDashboard";
+import StudentForm from "./views/StudentForm";
+import StudentsTable from "./views/StudentsTable";
+import StudentStats from "./views/StudentStats";
 import Testimonial from "./views/Testimonials";
 
 const ScrollToTop = () => {
@@ -23,6 +25,7 @@ const ScrollToTop = () => {
 function App() {
   return (
     <BrowserRouter>
+      <Navbar />
       <div className="App">
         <Route onChange={ScrollToTop} exact path="/" component={Home} />
 
@@ -101,6 +104,7 @@ function App() {
 
         <Route onChange={ScrollToTop} path="/projects" component={Projects} />
       </div>
+      <Footer />
     </BrowserRouter>
   );
 }
