+++ src/views/MentorStats.js
@@ -1,7 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { BACKEND_URL, STATS_API } from "../../constants";
-import Footer from "../Footer";
-import Navbar from "../Navbar";
+import { BACKEND_URL, STATS_API } from "../constants";
 
 export default function StudentDashboard() {
   const [fullName, setFullName] = useState("");
@@ -106,7 +104,6 @@ export default function StudentDashboard() {
              -> Important Announcements
              -> Stats of indiv Mentor ???
       */}
-        <Navbar className="is-black" />
         <div className="intro-card">
           <div className="avatar grow-card">
             <img
@@ -342,7 +339,6 @@ export default function StudentDashboard() {
           </div>
         </div> */}
       </div>
-      <Footer />
     </div>
   );
 }
